CD-2725: Split geostats and validator tests
